+++ ResourceDB/Entities/Models/bistro.scene.lua
@@ -3,7 +3,11 @@
 -- author @zoloypzuo
 return {
     type = "SCENE";
-    scene_file = "data/meshes/test.scene";
-    mesh_file = "data/meshes/test.meshes";
-    material_file = "data/meshes/test.materials";
+    --scene_file = "data/meshes/test.scene";
+    --mesh_file = "data/meshes/test.meshes";
+    --material_file = "data/meshes/test.materials";
+    scene_file = "data/meshes/bistro_all.meshes";
+    mesh_file = "data/meshes/bistro_all.scene";
+    material_file = "data/meshes/bistro_all.materials";
+    dummy_texture_file = "data/const1.bmp";
 }
